+++ resources/views/components/dropdown/header.blade.php
@@ -1,0 +1,4 @@
+<li {{ $attributes
+    ->class([
+        'dropdown-header',
+    ]) }}>{{ $slot }}</li>