+++ resources/views/components/nav/item.blade.php
@@ -8,16 +8,42 @@
     $itemAttributes = $attributes->whereDoesntStartWith('container-');
 
     $active = $attributes->get('href') === request()?->fullUrl();
+
+    /** @var ?\Illuminate\View\ComponentSlot $dropdown */
+    $hasDropdown = isset($dropdown);
+    if ($hasDropdown) {
+        $itemAttributes = $itemAttributes->merge([
+            'href' => $itemAttributes->get('href', '#'),
+            'role' => 'button',
+            'data-bs-toggle' => 'dropdown',
+            'aria-haspopup' => 'true',
+            'aria-expanded' => 'false',
+        ]);
+    }
 @endphp
-<li {{ $containerAttributes->class(['nav-item']) }}>
+<li {{ $containerAttributes
+    ->class([
+        'nav-item',
+        'dropdown' => $hasDropdown,
+    ]) }}>
     <a {{ $itemAttributes
         ->class([
             'nav-link',
             'active' => $active,
+            'dropdown-toggle' => $hasDropdown,
         ])
         ->merge([
             'aria-current' => $active ? 'page' : null,
             'aria-disabled' => $disabled ? 'true' : null,
             'disabled' => $disabled,
         ]) }}>{{ $slot }}</a>
+    @if($hasDropdown)
+        <ul {{ $dropdown->attributes
+            ->class([
+                'dropdown-menu',
+            ])
+            ->merge([
+                'aria-labelledby' => $itemAttributes->get('id'),
+            ]) }}>{{ $dropdown }}</ul>
+    @endif
 </li>