+++ CHANGELOG.md
@@ -12,5 +12,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Components for alerts, badges, and links
 - Components for breadcrumb and navigation
 - Components for buttons, groups, and toolbars
+- Components for dropdowns
 - Components for forms and their fields including feedback
 - Components for list groups and their items